Done all I can with the GUI
But it's not very good...
At the moment it renders a contact list with a few random examples in.
It can open a contact to view its details but it can't open more than
one contact. You may be able to get to the add contact screen but it
depends.